+++ components/Modules/Map/MapWindow.tsx
@@ -1,25 +1,21 @@
-import { PhotoMarkerIcon } from '@components/icons';
-import MapLayerMenu from '@components/Modules/Map/LayerMenu';
+import MapStyleMenu from '@components/Modules/Map/MapStyleMenu';
+import Marker from '@components/Modules/Map/Marker';
 import { SearchLocation } from '@components/Modules/SearchLocation';
 import { ConvertBounds } from '@lib/functions';
-import Viewport from '@lib/types/viewport';
-import Image from 'next/image';
+import GetMapStyle from '@lib/functions/map/GetMapStyle';
+import { IViewport, MapStyles } from '@lib/types/map';
 import { useTheme } from 'next-themes';
 import React, { useRef, useState } from 'react';
-import ReactMapGL, { ExtraState, MapRef, Marker } from 'react-map-gl';
+import ReactMapGL, { ExtraState, MapRef } from 'react-map-gl';
 import useSupercluster from 'use-supercluster';
 
 import { Maybe } from '../../../.cache/__types__';
-import { Dark, Light, Satellite } from '../../../shared/config/MapStyles';
-import UrlPrefix from '../../../shared/config/UrlPrefix';
-import { MapContext } from './MapContext';
+import { MapContext } from '../../Templates/Map/MapContext';
 
 const MapGL: React.FC = () => {
   const mapContext = React.useContext(MapContext);
 
-  const [mapStyle, setMapStyle] = React.useState<
-    'theme' | 'light' | 'dark' | 'satellite'
-  >('theme'); // possible map styles, defaults to theme
+  const [mapStyle, setMapStyle] = useState<MapStyles>('theme'); // possible map styles, defaults to theme
 
   const { resolvedTheme } = useTheme(); // get current theme
 
@@ -29,18 +25,7 @@ const MapGL: React.FC = () => {
     width: '100%',
     height: '100%',
     className: 'rounded-lg',
-    mapStyle:
-      // default map style by theme
-      mapStyle === 'theme'
-        ? resolvedTheme === 'dark'
-          ? Dark
-          : Light
-        : // explicit map styles
-        mapStyle === 'dark'
-        ? Dark
-        : mapStyle === 'satellite'
-        ? Satellite
-        : Light,
+    mapStyle: GetMapStyle(mapStyle, resolvedTheme),
     mapboxApiAccessToken: process.env.NEXT_PUBLIC_MAPBOX_TOKEN, // MAPBOX API ACCESS TOKEN
   };
 
@@ -94,7 +79,7 @@ const MapGL: React.FC = () => {
       <ReactMapGL
         {...mapContext.viewport}
         {...mapProps}
-        onViewportChange={(viewport: React.SetStateAction<Viewport>) =>
+        onViewportChange={(viewport: React.SetStateAction<IViewport>) =>
           mapContext.setViewport(viewport)
         }
         onInteractionStateChange={async (state: ExtraState) => {
@@ -126,7 +111,7 @@ const MapGL: React.FC = () => {
                       },
                     },
                   },
-                  updateQuery: (prev, { fetchMoreResult }) => {
+                  updateQuery: (prev: any, { fetchMoreResult }) => {
                     if (!fetchMoreResult) return prev; // if no new data, return prev data
                     const prevPlaces =
                       prev.places === undefined ||
@@ -138,7 +123,8 @@ const MapGL: React.FC = () => {
                       places: [
                         ...prevPlaces,
                         ...fetchMoreResult.places.filter(
-                          (place) => !prevPlaces.find((p) => p.id === place.id) // filter out duplicates
+                          (place: any) =>
+                            !prevPlaces.find((p: any) => p.id === place.id) // filter out duplicates
                         ),
                       ],
                     };
@@ -156,7 +142,7 @@ const MapGL: React.FC = () => {
                       },
                     },
                   },
-                  updateQuery: (prev, { fetchMoreResult }) => {
+                  updateQuery: (prev: any, { fetchMoreResult }) => {
                     if (!fetchMoreResult) return prev; // if no new data, return prev data
                     const prevPosts =
                       prev.posts.length === undefined ? [] : prev.posts;
@@ -165,7 +151,8 @@ const MapGL: React.FC = () => {
                       posts: [
                         ...prevPosts,
                         ...fetchMoreResult.posts.filter(
-                          (post) => !prevPosts.find((p) => p.id === post.id) // filter out duplicates
+                          (post: any) =>
+                            !prevPosts.find((p: any) => p.id === post.id) // filter out duplicates
                         ),
                       ],
                     };
@@ -187,7 +174,7 @@ const MapGL: React.FC = () => {
           );
 
           return (
-            <MapMarker
+            <Marker
               key={cluster.id}
               place={{
                 id: cluster.id,
@@ -224,106 +211,13 @@ const MapGL: React.FC = () => {
         })}
 
         <div className="absolute z-40 right-2 bottom-2">
-          <MapLayerMenu
+          <MapStyleMenu
             setMapStyle={setMapStyle}
             viewport={mapContext.viewport}
           />
         </div>
       </ReactMapGL>
     </>
-  );
-};
-
-type MapMarkerProps = {
-  place: {
-    id: number;
-    name?: string | null;
-    longitude: number;
-    latitude: number;
-    icon?: string | null;
-    preview: {
-      hash: string;
-    };
-  };
-  onClick?: () => void;
-  numberOfPlaces?: number;
-};
-
-const MapMarker: React.FC<MapMarkerProps> = ({
-  place,
-  onClick,
-  numberOfPlaces,
-}) => {
-  const mapContext = React.useContext(MapContext);
-  const [loadingImage, setLoadingImage] = useState<boolean>(true);
-
-  return (
-    <Marker
-      key={place.id}
-      latitude={place.latitude}
-      longitude={place.longitude}
-    >
-      <div className="relative">
-        <div
-          className={`absolute -translate-x-1/2 -translate-y-1/2 cursor-pointer rounded-full border-2 
-            ${place.icon ? 'w-24 h-24' : 'w-16 h-16'}
-            ${
-              !place.icon && !loadingImage
-                ? 'border-white hover:border-brand'
-                : 'border-transparent'
-            }`}
-          onMouseEnter={() => mapContext.setHoverPlaceId(place.id)}
-          onMouseLeave={() => mapContext.setHoverPlaceId(null)}
-          onClick={onClick}
-        >
-          {place.icon ? (
-            // if place has an icon
-            <Image
-              src={UrlPrefix + place.icon}
-              width={90}
-              height={90}
-              objectFit="contain"
-              alt="Picture on map"
-              quality={10}
-            />
-          ) : place.preview.hash ? (
-            // else show preview image of place
-            <>
-              {loadingImage && ( // if image is loading show universal place icon
-                <div className="p-4 text-brand">
-                  <PhotoMarkerIcon />
-                </div>
-              )}
-              {numberOfPlaces !== undefined &&
-                numberOfPlaces > 1 && ( // if marker is a cluster
-                  <div
-                    className={`absolute ${
-                      loadingImage ? 'top-0 right-0' : '-top-2 -right-2'
-                    } w-6 h-6 text-white text-center bg-brand rounded-full z-20`}
-                  >
-                    {numberOfPlaces}
-                  </div>
-                )}
-
-              <Image
-                src={UrlPrefix + place.preview.hash}
-                placeholder="empty"
-                layout="fill"
-                onLoadingComplete={() => setLoadingImage(false)}
-                objectFit="cover"
-                objectPosition="center"
-                className="p-4 rounded-full"
-                alt="Picture on map"
-              />
-            </>
-          ) : (
-            <div className="p-4 text-brand">
-              <PhotoMarkerIcon />
-            </div>
-          )}
-        </div>
-      </div>
-    </Marker>
   );
 };
