+++ lib/functions/index.ts
@@ -1,6 +1,6 @@
-export { default as ConvertBounds } from './ConvertBounds';
 export { default as FirstLetterUppercase } from './FirstLetterUppercase';
 export { default as GetCookieFromServerSideProps } from './GetCookieFromServerSideProps';
 export { default as IsJwtValid } from './IsJwtValid';
 export { default as LogOut } from './LogOut';
+export { default as ConvertBounds } from './map/ConvertBounds';
 export { default as SSRRedirect } from './SSRRedirect';
