--- lib/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as hoverHandler } from './hoverHandler';
