+++ lib/types/forms/forgotPasswordFormInputs.ts
@@ -1,5 +1,5 @@
-type ForgotPasswordFormInputs = {
+interface ForgotPasswordFormInputs {
   login: string;
-};
+}
 
 export default ForgotPasswordFormInputs;
