+++ lib/types/forms/loginFormInputs.ts
@@ -1,6 +1,6 @@
-type LoginFormInputs = {
+interface LoginFormInputs {
   login: string;
   password: string;
-};
+}
 
 export default LoginFormInputs;
