+++ lib/types/forms/newPasswordFormInputs.ts
@@ -1,6 +1,6 @@
-type NewPasswordFormInputs = {
+interface NewPasswordFormInputs {
   password: string;
   repeatPassword: string;
-};
+}
 
 export default NewPasswordFormInputs;
