--- lib/types/viewport.ts
@@ -1,11 +1,0 @@
-// react-map-gl type
-
-type Viewport = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-  bearing: number;
-  pitch: number;
-};
-
-export default Viewport;
