--- types/Bounds.ts
@@ -1,8 +1,0 @@
-type Bounds = {
-  maxLatitude: number;
-  minLatitude: number;
-  maxLongitude: number;
-  minLongitude: number;
-};
-
-export default Bounds;
